refactor: useInlineEdit フック抽出 + EdgeContextMenu コンポーネント分離
- useInlineEdit フックを新規作成し, EditableNode / GroupNode /
  EditableLabelEdge に散在していた editing / inputValue / composing /
  cancelledRef / startEdit / confirm / cancel の重複ロジックを一元化
- EdgeContextMenu を GraphEditor.tsx から独立コンポーネントとして抽出

Confidence: high
Scope-risk: narrow
