+++ src/client/src/GraphEditor.tsx
@@ -18,7 +18,8 @@ import {
 } from '@xyflow/react';
 import { useCallback, useEffect, useMemo, useRef } from 'react';
 import '@xyflow/react/dist/style.css';
-import type { EdgePathType, GraphFile } from '@conversensus/shared';
+import type { GraphFile } from '@conversensus/shared';
+import { EdgeContextMenu } from './EdgeContextMenu';
 import { EditableLabelEdge } from './EditableLabelEdge';
 import { EditableNode } from './EditableNode';
 import { GroupNode } from './GroupNode';
@@ -200,71 +201,7 @@ function GraphEditorInner({ file, onChange }: Props) {
         </Panel>
       </ReactFlow>
       {contextMenu && (
-        // biome-ignore lint/a11y/noStaticElementInteractions: context menu uses mousedown to block propagation
-        <div
-          style={{
-            position: 'fixed',
-            top: contextMenu.y,
-            left: contextMenu.x,
-            background: '#fff',
-            border: '1px solid #ddd',
-            borderRadius: 6,
-            boxShadow: '0 2px 8px rgba(0,0,0,0.15)',
-            zIndex: 1000,
-            minWidth: 160,
-            padding: '4px 0',
-          }}
-          onMouseDown={(e) => e.stopPropagation()}
-        >
-          <div
-            style={{
-              padding: '4px 14px 6px',
-              fontSize: 11,
-              color: '#888',
-              borderBottom: '1px solid #eee',
-              marginBottom: 4,
-            }}
-          >
-            {contextMenu.targetEdgeIds.length === 1
-              ? 'エッジの種類'
-              : `${contextMenu.targetEdgeIds.length} 本のエッジを変更`}
-          </div>
-          {(
-            [
-              ['bezier', 'Bezier（曲線）'],
-              ['straight', 'Straight（直線）'],
-              ['step', 'Step（直角）'],
-              ['smoothstep', 'Smooth Step（角丸）'],
-            ] as [EdgePathType, string][]
-          ).map(([type, label]) => {
-            const isCurrent = contextMenu.currentPathType === type;
-            return (
-              <button
-                key={type}
-                type="button"
-                onClick={() => setEdgePathType(contextMenu.targetEdgeIds, type)}
-                style={{
-                  display: 'flex',
-                  alignItems: 'center',
-                  gap: 6,
-                  width: '100%',
-                  padding: '6px 14px',
-                  textAlign: 'left',
-                  background: 'none',
-                  border: 'none',
-                  fontSize: 13,
-                  fontWeight: isCurrent ? 'bold' : 'normal',
-                  cursor: 'pointer',
-                }}
-              >
-                <span style={{ width: 12, flexShrink: 0 }}>
-                  {isCurrent ? '✓' : ''}
-                </span>
-                {label}
-              </button>
-            );
-          })}
-        </div>
+        <EdgeContextMenu contextMenu={contextMenu} onSelect={setEdgePathType} />
       )}
     </div>
   );
